+++ src/packets/data/WorldPosData.h
@@ -8,22 +8,22 @@ class Packet;
 class WorldPosData
 {
 public:
-	float x;
-	float y;
+	double x;
+	double y;
 
 	WorldPosData() {};
 	WorldPosData(Packet*);
 	WorldPosData(const WorldPosData&);
-	WorldPosData(float, float); // {x,y}
+	WorldPosData(double, double); // {x,y}
 
 	void Read(Packet*);
 	void Write(Packet*);
 
 	bool outOfBounds(int) const;
 
-	float distanceTo(WorldPosData&) const;
-	float sqDistanceTo(WorldPosData&) const;
-	float angleTo(WorldPosData&) const;
+	double distanceTo(WorldPosData&) const;
+	double sqDistanceTo(WorldPosData&) const;
+	double angleTo(WorldPosData&) const;
 
 	bool operator==(const WorldPosData&) const;
 	bool operator!=(const WorldPosData&) const;
